requirements.txt
playwright==1.42.0
pandas==2.2.1
openpyxl==3.1.2
python-dotenv